// src/types/api.d.ts
export interface components {
    schemas: {
        RegisterDto: {
            name: string;
            /** Format: email */
            email: string;
            password: string;
        };
        AuthResponseDto: {
            accessToken: string;
            user: {
                /** Format: uuid */
                id: string;
                /** Format: email */
                email: string;
                name: string | null;
                /** @enum {string} */
                role: "admin" | "editor";
            };
        };
        LoginDto: {
            /** Format: email */
            email: string;
            password: string;
        };
        UserResponseDto: {
            /** Format: uuid */
            id: string;
            /** Format: email */
            email: string;
            name: string | null;
            /** @enum {string} */
            role: "admin" | "editor";
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}